updated specs
